+++ src/imports/api/accounts/emailTemplates/resetPassword.js
@@ -1,6 +1,9 @@
 import { i18n } from '../../language/language'
 import { Meteor } from 'meteor/meteor'
 import { getCredentialsAsBuffer, getFullName } from './common'
+import {createLog} from '../../log/createLog'
+
+const debug = createLog({ name: 'email/resetPassword', type: 'debug' })
 
 export const getResetPasswordSubject = ({ siteName, defaultLocale }) => user => {
   const locale = user?.locale || defaultLocale
@@ -21,8 +24,7 @@ export const getResetPasswordText = ({ expiration, defaultLocale, supportEmail }
   const text = i18n.get(locale, 'accounts.resetPassword.text', textOptions)
 
   if (Meteor.isDevelopment && !Meteor.isTest) {
-    console.debug(textOptions.url)
-    console.debug(text)
+    debug('body', textOptions.url, text)
   }
 
   return text
