+++ src/imports/api/accounts/emailTemplates/verifyEmail.js
@@ -1,6 +1,9 @@
 import { Meteor } from 'meteor/meteor'
 import { i18n } from '../../language/language'
 import { getFullName } from './common'
+import {createLog} from '../../log/createLog'
+
+const debug = createLog({ name: 'email/verify', type: 'debug' })
 
 export const getVerifyEmailSubject = ({ siteName, defaultLocale }) => user => {
   const locale = user?.locale || defaultLocale
@@ -15,8 +18,7 @@ export const getVeryFyEmailText = ({ defaultLocale, supportEmail }) => (user, or
   const text = i18n.get(locale, 'accounts.verifyEmail.text', { name, url, supportEmail })
 
   if (Meteor.isDevelopment && !Meteor.isTest) {
-    console.debug(url)
-    console.debug(text)
+    debug(url, text)
   }
 
   return text
