+++ src/imports/api/accounts/registration/UserFactory.js
@@ -5,14 +5,36 @@ import { Schema } from '../../schema/Schema'
 import { Roles } from 'meteor/alanning:roles'
 import { rollbackAccount } from './rollbackAccount'
 import { userExists } from '../user/userExists'
+import {createLog} from '../../log/createLog'
 
-let createSchema
 
+/**
+ * Creates new user accounts
+ * @namespace
+ */
 export const UserFactory = {}
 
 UserFactory.name = 'UserFactory'
 
+const debug = createLog({ name: UserFactory.name, type: 'debug'})
+let createSchema
+
+/**
+ * Creates a new user account by given options. Args are validated.
+ * @param email {string}
+ * @param password {string}
+ * @param role {string}
+ * @param firstName {string}
+ * @param lastName {string}
+ * @param institution {string}
+ * @param locale {string=}
+ * @returns {string} the new user's document _id
+ * @throws {Meteor.Error} if user exists by given Email
+ * @throws {Meteor.Error} if user has not been created
+ * @throws {Meteor.Error} if user has not successfully been assigned to given roles
+ */
 UserFactory.create = function create ({ email, password, role, firstName, lastName, institution, locale }) {
+  debug('create new user', { email, institution, role })
   if (!createSchema) {
     createSchema = Schema.create(createUserSchema)
   }
@@ -64,7 +86,7 @@ UserFactory.create = function create ({ email, password, role, firstName, lastNa
     throw new Meteor.Error('createUser.failed', 'createUser.profileNotUpdated', email)
   }
 
-  console.debug('add user to roles', userId, [role], institution)
+  debug('add user to roles', userId, [role], institution)
   // adds the user to the given roles and scope
   Roles.addUsersToRoles(userId, [role], institution)
   if (!Roles.userIsInRole(userId, [role], institution)) {
