+++ src/imports/api/config/Features.js
@@ -1,6 +1,7 @@
 import { Meteor } from 'meteor/meteor'
 
 const features = Object.create(null)
+
 Object.assign(features, Meteor.settings.public.features)
 
 /**
@@ -10,8 +11,7 @@ export const Features = {}
 
 Features.get = (name) => {
   if (!name || !Object.hasOwnProperty.call(features, name)) {
-    console.debug(name, features, features[name])
-    throw new Error(`Features has no feature by name ${name}`)
+    throw new Error(`Features have no feature by name ${name}`)
   }
   return features[name]
 }
