+++ src/imports/api/utils/insertUpdate.js
@@ -1,4 +1,7 @@
 import crypto from 'crypto'
+import {createLog} from '../log/createLog'
+
+const debugLog = createLog({ name: 'insertUpdate', type: 'debug' })
 
 const sortedProps = (doc) => {
   let obj = ''
@@ -19,10 +22,12 @@ const hash = (doc = {}) => {
 const areEqual = (doc1, doc2, debug) => {
   const h1 = hash(doc1)
   const h2 = hash(doc2)
+
   if (debug && h1 !== h2) {
-    console.info(`Hash 1: ${h1}`)
-    console.info(`Hash 2: ${h2}`)
+    debugLog(`Hash 1: ${h1}`)
+    debugLog(`Hash 2: ${h2}`)
   }
+
   return h1 === h2
 }
 
@@ -44,7 +49,7 @@ export const insertUpdate = function insertUpdate (collection, doc, hash, debug)
     if (!existingDoc) {
       const insertDocId = collection.insert(doc)
       if (debug) {
-        console.info(`[${collection._name}]: insert ${insertDocId}`)
+        debugLog(`[${collection._name}]: insert ${insertDocId}`)
       }
       return insertDocId
     }
@@ -59,14 +64,11 @@ export const insertUpdate = function insertUpdate (collection, doc, hash, debug)
     delete doc._id
     const updated = collection.update(docId, { $set: doc })
     if (debug) {
-      console.info(`[${collection._name}]: update ${docId} ${updated}`)
+      debugLog(`[${collection._name}]: update ${docId} ${updated}`)
     }
     return updated
   }
   catch (e) {
-    console.log(e.message)
-    console.log('collection: ' + collection._name)
-    console.log('document: ')
-    console.dir(doc)
+    console.error(e)
   }
 }
