+++ src/imports/contexts/beamer/Beamer.js
@@ -185,14 +185,16 @@ Beamer.methods.insert = {
   timeInterval: 50000,
   run: onServer(function () {
     const BeamerCollection = getCollection(Beamer.name)
-    console.log(BeamerCollection.findOne({ createdBy: this.userId }))
+
     if (BeamerCollection.findOne({ createdBy: this.userId })) {
       throw new Meteor.Error('errors.docAlreadyExists')
     }
+
     const ui = {
       background: Beamer.defaultBackground,
       grid: Beamer.defaultGridlayout
     }
+
     return BeamerCollection.insert({ createdBy: this.userId, references: [], ui })
   })
 }
