+++ src/imports/contexts/classroom/group/Group.js
@@ -158,9 +158,7 @@ Group.publications.my = {
 
     query.$or.push(myGroups, iamMember)
 
-    const cursor = getCollection(Group.name).find(query, { fields: Group.publicFields })
-    console.log(this.userId, JSON.stringify(query), cursor.count())
-    return cursor
+    return getCollection(Group.name).find(query, { fields: Group.publicFields })
   })
 }
 
@@ -174,9 +172,7 @@ Group.publications.single = {
   run: onServer(function ({ groupId }) {
     const { userId } = this
     const query = { _id: groupId, users: { $elemMatch: { userId } } }
-    const cursor = getCollection(Group.name).find(query, { fields: Group.publicFields })
-    console.log(this.userId, JSON.stringify(query), cursor.count())
-    return cursor
+    return getCollection(Group.name).find(query, { fields: Group.publicFields })
   })
 }
