+++ src/imports/contexts/system/accounts/users/methods/setResearch.js
@@ -18,7 +18,6 @@ export const setResearch = function setResearch ({ participate }) {
 
   if (participate) {
     const user = Meteor.users.findOne(userId)
-    console.debug(user)
     const { firstName, lastName } = user
     const token = createResearchConfirmToken({ userId })
     const email = user.emails[0].address
